fix decomposition webp, webp format fixing, error fixing

+++ internal/repository/images/webp.go
@@ -1,6 +1,7 @@
-package repository
+package images
 
 import (
+	"HeadHunter/internal/errorHandler"
 	"github.com/kolesa-team/go-webp/encoder"
 	"github.com/kolesa-team/go-webp/webp"
 	"image"
@@ -9,6 +10,10 @@ import (
 )
 
 func UploadWebpImage(path, name string, image *image.Image) (err error) {
+	if name == "" || path == "" {
+		return errorHandler.ErrBadRequest
+	}
+
 	resultImage, createErr := os.Create(strings.Join([]string{path, name}, ""))
 	if createErr != nil {
 		return createErr
@@ -32,6 +37,10 @@ func UploadWebpImage(path, name string, image *image.Image) (err error) {
 }
 
 func DeleteWebpImage(path, name string) error {
+	if path == "" || name == "" {
+		return errorHandler.ErrBadRequest
+	}
+
 	removeErr := os.Remove(strings.Join([]string{path, name}, ""))
 	if removeErr != nil {
 		return removeErr
